tests: 64 bit compatibility
Fixed compilation errors. Mostly DEBUG/printf formatting and void pointer casting.

Added another hardcoded offset for gnrc_sixlowpan_frag_* tests.
Increased mem size of lua test in case of 64 bit.
Reduced required backtrace size to 3 in native backtrace test.

+++ tests/sys/shell/main.c
@@ -87,7 +87,7 @@ static int print_shell_bufsize(int argc, char **argv)
 {
     (void)argc;
     (void)argv;
-    printf("%d\n", sizeof(line_buf));
+    printf("%u\n", (unsigned)sizeof(line_buf));
 
     return 0;
 }
